fix(backend/api): delete db config file, delete useless migrations files, reorganize src folder
