+++ apps/backend/api/src/app/app.module.ts
@@ -1,8 +1,7 @@
-// app.module.ts
 import { Module } from '@nestjs/common';
-import { entities } from './config/entities';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { entities } from '../config/entities';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UsersModule } from './users/users.module';
@@ -11,6 +10,9 @@ import { GamesModule } from './games/games.module';
 
 @Module({
   imports: [
+    ConfigModule.forRoot({
+      isGlobal: true,
+    }),
     TypeOrmModule.forRootAsync({
       imports: [ConfigModule],
       useFactory: (configService: ConfigService) => ({
