+++ apps/backend/api/src/app/characters/character.service.spec.ts
