+++ apps/backend/api/src/app/characters/entities/character.entity.ts
