+++ apps/backend/api/src/app/rooms/entities/room.entity.ts
