+++ apps/backend/api/src/app/rooms/entities/room.spec.ts
