+++ apps/backend/api/src/app/rooms/rooms.service.spec.ts
