+++ apps/backend/api/src/app/rooms/rooms.service.ts
