+++ apps/backend/api/src/app/users/entities/user.entity.ts
