--- apps/backend/api/src/config/database.config.ts
@@ -1,13 +1,0 @@
-import { DataSourceOptions } from 'typeorm';
-import { entities } from './entities';
-
-export const databaseConfig: DataSourceOptions = {
-  type: 'mysql',
-  host: process.env.DATABASE_HOST,
-  port: parseInt(process.env.DATABASE_PORT || '3000'),
-  username: process.env.DATABASE_USERNAME,
-  password: process.env.DATABASE_PASSWORD,
-  database: process.env.DATABASE_DB,
-  entities: entities,
-  synchronize: false,
-};
