+++ apps/backend/api/src/config/entities.ts
@@ -1,6 +1,6 @@
-import { Character } from '../characters/entities/character.entity';
-import { Game } from '../games/entities/game.entity';
-import { User } from '../users/entities/user.entity';
-import { Room } from '../rooms/room/room.entity';
+import { Character } from '../app/characters/entities/character.entity';
+import { Game } from '../app/games/entities/game.entity';
+import { User } from '../app/users/entities/user.entity';
+import { Room } from '../app/rooms/entities/room.entity';
 
 export const entities = [Character, Game, User, Room];
